# Replace `getproperty` parameter data access with typed `getindex` to improve inference