--- a/src/C3D.jl
+++ b/src/C3D.jl
@@ -79,18 +79,18 @@ function calcresiduals(x::AbstractVector, scale)
     residuals = (reinterpret.(UInt32, x) .>> 16) .& 0xff .* scale
 end
 
-function readdata(f::IOStream, groups::Dict{Symbol,Group}, FEND::Endian, FType::Type{T}) where T <: Union{Float32,VaxFloatF}
-    seek(f, (groups[:POINT].DATA_START-1)*512)
+function readdata(io::IOStream, groups::Dict{Symbol,Group}, FEND::Endian, FType::Type{T}) where T <: Union{Float32,VaxFloatF}
+    seek(io, (groups[:POINT].DATA_START-1)*512)
 
     format = groups[:POINT].SCALE > 0 ? Int16 : FType
 
     # Read data in a transposed structure for better read/write speeds due to Julia being
     # column-order arrays
-    numframes = convert(Int, groups[:POINT].FRAMES)
+    numframes::Int = groups[:POINT].FRAMES
     if numframes == typemax(UInt16) && haskey(groups, :TRIAL) && haskey(groups[:TRIAL].params, :ACTUAL_END_FIELD)
         numframes = convert(Int, reinterpret(Int32, groups[:TRIAL].ACTUAL_END_FIELD)[1])
     end
-    nummarkers = convert(Int, groups[:POINT].USED)
+    nummarkers::Int = groups[:POINT].USED
     hasmarkers = !iszero(nummarkers)
     if hasmarkers
         point = Array{Float32,2}(undef, nummarkers*3, numframes)
@@ -100,7 +100,7 @@ function readdata(f::IOStream, groups::Dict{Symbol,Group}, FEND::Endian, FType::
         pointidxs = filter(x -> x % 4 != 0, 1:nb)
         residxs = filter(x -> x % 4 == 0, 1:nb)
 
-        pointtmp = Array{format}(undef, nb)
+        pointtmp = Vector{format}(undef, nb)
         pointview = view(pointtmp, pointidxs)
         resview = view(pointtmp, residxs)
     else
@@ -108,43 +108,48 @@ function readdata(f::IOStream, groups::Dict{Symbol,Group}, FEND::Endian, FType::
         residuals = Array{Float32,2}(undef, 0,0)
     end
 
-    numchannels = convert(Int, groups[:ANALOG].USED)
+    numchannels::Int = groups[:ANALOG].USED
     haschannels = !iszero(numchannels)
     if haschannels
         # Analog Samples Per Frame => ASPF
-        aspf = convert(Int, groups[:ANALOG].RATE/groups[:POINT].RATE)
+        aspf::Int = groups[:ANALOG].RATE/groups[:POINT].RATE
         analog = Array{Float32,2}(undef, numchannels, aspf*numframes)
 
-        analogtmp = Array{format}(undef, (numchannels,aspf))
+        analogtmp = Matrix{format}(undef, (numchannels,aspf))
     else
         analog = Array{Float32,2}(undef, 0,0)
     end
 
     @inbounds for i in 1:numframes
         if hasmarkers
-            saferead!(f, pointtmp, FEND)
-            point[:,i] = convert.(Float32, pointview) # Convert from `format` (eg Int16 or FType)
-            residuals[:,i] = convert.(Float32, resview)
+            saferead!(io, pointtmp, FEND)
+            point[:,i] = pointview
+            residuals[:,i] = resview
         end
         if haschannels
-            saferead!(f, analogtmp, FEND)
-            analog[:,((i-1)*aspf+1):(i*aspf)] = convert.(Float32, analogtmp)
+            saferead!(io, analogtmp, FEND)
+            analog[:,((i-1)*aspf+1):(i*aspf)] = analogtmp
         end
     end
 
     if hasmarkers && format == Int16
         # Multiply or divide by [:point][:scale]
-        point .*= abs(groups[:POINT].SCALE)
+        POINT_SCALE::Float32 = groups[:POINT].SCALE
+        point .*= abs(POINT_SCALE)
     end
 
     if haschannels
         if numchannels == 1
-            analog[:] = (analog .- groups[:ANALOG].OFFSET) .*
-                        (groups[:ANALOG].GEN_SCALE * groups[:ANALOG].SCALE)
+            ANALOG_OFFSET = groups[:ANALOG].OFFSET::Float32
+            SCALE = (groups[:ANALOG].GEN_SCALE * groups[:ANALOG].SCALE)::Float32
+            analog .= (analog .- ANALOG_OFFSET) .* SCALE
         else
-            analog[:] = (analog .- groups[:ANALOG].OFFSET[1:numchannels]) .*
-                        groups[:ANALOG].GEN_SCALE .*
-                        groups[:ANALOG].SCALE[1:numchannels]
+            VECANALOG_OFFSET = convert(Vector{Float32},
+                groups[:ANALOG].OFFSET[1:numchannels])::Vector{Float32}
+            VECSCALE = convert(Vector{Float32}, groups[:ANALOG].GEN_SCALE .*
+                               groups[:ANALOG].SCALE[1:numchannels])::Vector{Float32}
+
+            analog .= (analog .- VECANALOG_OFFSET) .* VECSCALE
         end
     end
 
@@ -184,11 +189,11 @@ function saferead(io::IOStream, ::Type{VaxFloatF}, FEND::Endian)::Float32
     end
 end
 
-function saferead(io::IOStream, ::Type{VaxFloatF}, FEND::Endian, dims)::Array{Float32}
+function saferead(io::IOStream, ::Type{VaxFloatF}, FEND::Endian, dims::NTuple{N,Int})::Array{Float32,N} where N
     if FEND == LE
-        return convert.(Float32, ltoh.(read!(io, Array{VaxFloatF}(undef, dims))))
+        return convert(Array{Float32,N}, ltoh.(read!(io, Array{VaxFloatF}(undef, dims))))
     else
-        return convert.(Float32, ntoh.(read!(io, Array{VaxFloatF}(undef, dims))))
+        return convert(Array{Float32,N}, ntoh.(read!(io, Array{VaxFloatF}(undef, dims))))
     end
 end
 
@@ -208,12 +213,12 @@ function readc3d(fn::AbstractString; paramsonly=false, validate=true,
         error("File ", fn, " cannot be found")
     end
 
-    f = open(fn, "r")
+    io = open(fn, "r")
 
-    groups, header, FEND, FType = _readparams(fn, f)
+    groups, header, FEND, FType = _readparams(fn, io)
 
     if validate
-        validatec3d(header, groups, complete=false)
+        validatec3d(header, groups)
     end
 
     if paramsonly
@@ -221,32 +226,32 @@ function readc3d(fn::AbstractString; paramsonly=false, validate=true,
         residual = Dict{String,Array{Union{Missing, Float32},1}}()
         analog = Dict{String,Array{Float32,1}}()
     else
-        (point, residual, analog) = readdata(f, groups, FEND, FType)
+        (point, residual, analog) = readdata(io, groups, FEND, FType)
     end
 
-    res = C3DFile(fn, header, groups, point, residual, analog; missingpoints=missingpoints)
+    close(io)
 
-    close(f)
+    res = C3DFile(fn, header, groups, point, residual, analog; missingpoints=missingpoints)
 
     return res
 end
 
-function _readparams(fn::String, f::IOStream)
-    params_ptr = read(f, UInt8)
+function _readparams(fn::String, io::IOStream)
+    params_ptr = read(io, UInt8)
 
-    if read(f, UInt8) != 0x50
+    if read(io, UInt8) != 0x50
         error("File ", fn, " is not a valid C3D file")
     end
 
     # Jump to parameters block
-    seek(f, (params_ptr - 1) * 512)
+    seek(io, (params_ptr - 1) * 512)
 
     # Skip 2 reserved bytes
     # TODO: store bytes for saving modified files
-    skip(f, 2)
+    skip(io, 2)
 
-    paramblocks = read(f, UInt8)
-    proctype = read(f, Int8) - 83
+    paramblocks = read(io, UInt8)
+    proctype = read(io, Int8) - 83
 
     FType = Float32
 
@@ -265,89 +270,89 @@ function _readparams(fn::String, f::IOStream)
         error("Malformed processor type. Expected 1, 2, or 3. Found ", proctype)
     end
 
-    mark(f)
-    header = readheader(f, FEND, FType)
-    reset(f)
-    unmark(f)
+    mark(io)
+    header = readheader(io, FEND, FType)
+    reset(io)
+    unmark(io)
 
     gs = Array{Group,1}()
-    ps = Array{AbstractParameter,1}()
+    ps = Array{Parameter,1}()
     moreparams = true
     fail = 0
     np = 0
 
-    skip(f, 1)
-    if read(f, Int8) < 0
+    skip(io, 1)
+    if read(io, Int8) < 0
         # Group
-        skip(f, -2)
-        push!(gs, readgroup(f, FEND, FType))
+        skip(io, -2)
+        push!(gs, readgroup(io, FEND, FType))
         np = gs[end].pos + gs[end].np + abs(gs[end].nl) + 2
         moreparams = gs[end].np != 0 ? true : false
     else
         # Parameter
-        skip(f, -2)
-        push!(ps, readparam(f, FEND, FType))
-        np = ps[end].pos + ps[end].np + abs(ps[end].nl) + 2
-        moreparams = ps[end].np != 0 ? true : false
+        skip(io, -2)
+        push!(ps, readparam(io, FEND, FType))
+        np = ps[end].pos::Int + ps[end].np::Int16 + abs(ps[end].nl::Int8) + 2
+        moreparams = ps[end].np::Int16 != 0 ? true : false
     end
 
     while moreparams
         # Mark current position in file in case the pointer is incorrect
-        mark(f)
-        if fail === 0 && np != position(f)
-                @debug "Pointer mismatch at position $(position(f)) where pointer was $np"
-                seek(f, np)
+        mark(io)
+        if fail === 0 && np != position(io)
+                # @debug "Pointer mismatch at position $(position(io)) where pointer was $np"
+                seek(io, np)
         elseif fail > 1 # this is the second failed attempt
-            @debug "Second failed parameter read attempt from $(position(f))"
+            # @debug "Second failed parameter read attempt from $(position(io))"
             break
         end
 
         # Read the next two bytes to get the gid
-        skip(f, 1)
-        local gid = read(f, Int8)
+        skip(io, 1)
+        local gid = read(io, Int8)
         if gid < 0 # Group
             # Reset to the beginning of the group
-            skip(f, -2)
+            skip(io, -2)
             try
-                push!(gs, readgroup(f, FEND, FType))
+                push!(gs, readgroup(io, FEND, FType))
                 np = gs[end].pos + gs[end].np + abs(gs[end].nl) + 2
                 moreparams = gs[end].np != 0 ? true : false # break if the pointer is 0 (ie the parameters are finished)
                 fail = 0 # reset fail counter following a successful read
             catch e
                 # Last readgroup failed, possibly due to a bad pointer. Reset to the ending
                 # location of the last successfully read parameter and try again. Count the failure.
-                reset(f)
-                @debug "Read group failed, last parameter ended at $(position(f)), pointer at $np" fail
+                reset(io)
+                # @debug "Read group failed, last parameter ended at $(position(io)), pointer at $np" fail
                 fail += 1
             finally
-                unmark(f) # Unmark the file regardless
+                unmark(io) # Unmark the file regardless
             end
         elseif gid > 0 # Parameter
             # Reset to the beginning of the parameter
-            skip(f, -2)
+            skip(io, -2)
             try
-                push!(ps, readparam(f, FEND, FType))
-                np = ps[end].pos + ps[end].np + abs(ps[end].nl) + 2
-                moreparams = ps[end].np != 0 ? true : false
+                push!(ps, readparam(io, FEND, FType))
+                np = ps[end].pos::Int + ps[end].np::Int16 + abs(ps[end].nl::Int8) + 2
+                moreparams = ps[end].np::Int16 != 0 ? true : false
                 fail = 0
             catch e
-                reset(f)
-                @debug "Read group failed, last parameter ended at $(position(f)), pointer at $np" fail
+                reset(io)
+                # @debug "Read group failed, last parameter ended at $(position(io)), pointer at $np" fail
                 fail += 1
             finally
-                unmark(f)
+                unmark(io)
             end
         else # Last parameter pointer is incorrect (assumption)
             # The group ID should never be zero, if it is, the most likely explanation is
             # that the pointer is incorrect (eg the pointer was not fixed when the previously
             # last parameter was deleted or moved)
-            @debug "Bad last position. Assuming parameter section is finished."
+            # @debug "Bad last position. Assuming parameter section is finished."
             break
         end
     end
 
     groups = Dict{Symbol,Group}()
-    gids = Dict{Int,Symbol}()
+    gids = Dict{Int8,Symbol}()
 
     for group in gs
         groups[group.symname] = group